Require integer BL quantities

+++ backend/IQBF.Application/DTOs/BLs/CreateBLRequest.cs
@@ -1,2 +1,2 @@
 namespace IQBF.Application.DTOs.BLs;
-public sealed record CreateBLRequest(string Code, decimal TotalQuantity, Guid ShipId, Guid ProductId);
+public sealed record CreateBLRequest(string Code, int TotalQuantity, Guid ShipId, Guid ProductId);
